+++ rex/data/collate_fn.py
@@ -30,13 +30,13 @@ class GeneralCollateFn(object):
         self,
         key2type: Optional[Dict[str, Any]],
         guessing: Optional[bool] = False,
-        none_as_missing_keys: Optional[bool] = False,
+        missing_key_as_null: Optional[bool] = False,
     ) -> None:
         self.key2type = {}
         if isinstance(key2type, dict) and len(key2type) > 0:
             self.key2type.update(key2type)
         self.guessing = guessing
-        self.none_as_missing_keys = none_as_missing_keys
+        self.missing_key_as_null = missing_key_as_null
 
     def update_type_mapping(self, key2type: dict):
         for key, val_type in key2type.items():
@@ -75,7 +75,7 @@ class GeneralCollateFn(object):
         if len(self.key2type) == 0 and self.guessing:
             self.guess_types(data[0], update=True)
 
-        if self.none_as_missing_keys:
+        if self.missing_key_as_null:
             for d in data:
                 missing_keys = set(d.keys()) - set(self.key2type.keys())
                 for mkey in missing_keys: